Conflicts:
	flow/alocpro-mock-client/pom.xml
	flow/core/pom.xml
	flow/persistence/pom.xml
	flow/pom.xml
	flow/rest-client/pom.xml
	flow/schedule/pom.xml
	flow/test/pom.xml
	form/core/pom.xml
	form/exemplos-form/pom.xml
	form/pom.xml
	form/service/pom.xml
	form/ui-showcase/pom.xml
	form/wicket/pom.xml
	lib/commons/pom.xml
	lib/pom.xml
	lib/singular-support/pom.xml
	lib/wicket-utils/pom.xml
	pom.xml
	resources/pom.xml
	resources/ui-static-resources/pom.xml
	ws/pom.xml
	ws/wkhtmltopdf-client/pom.xml
	ws/wkhtmltopdf-ws/pom.xml

+++ form/core/src/main/java/org/opensingular/form/type/core/SILong.java
@@ -22,6 +22,6 @@ public class SILong extends SINumber<Long> implements SIComparable<Long> {
     }
 
     public Long getLong() {
-        return (Long) getValue();
+        return getValue();
     }
 }
